@@ -15,14 +15,14 @@
     "@types/csurf": "^1.9.36",
     "@types/express": "^4.17.2",
     "@types/express-session": "^1.15.16",
-    "@types/node": "^13.7.0",
-    "@types/nodemailer": "^6.4.0",
+    "@types/node": "^13.7.6",
     "@types/pg": "^7.14.1",
-    "@types/uuid": "^3.4.7",
+    "@types/uuid": "^3.4.8",
     "ts-node": "^8.6.2",
-    "typescript": "^3.7.5"
+    "typescript": "^3.8.2"
   },
   "dependencies": {
+    "@sendgrid/mail": "^6.5.3",
     "bcryptjs": "^2.4.3",
     "connect-flash": "^0.1.1",
     "connect-pg-simple": "^6.1.0",
@@ -31,9 +31,7 @@
     "ejs": "^3.0.1",
     "express": "^4.17.1",
     "express-session": "^1.17.0",
-    "nodemailer": "^6.4.3",
-    "nodemailer-sendgrid-transport": "^0.2.0",
-    "pg": "^7.18.1",
-    "uuid": "^7.0.0"
+    "pg": "^7.18.2",
+    "uuid": "^7.0.1"
   }
 }
